chore(client): refactor location save, add db lookup first
