+++ client/actions/index.js
@@ -1,2 +1,3 @@
 export * from './events'
+export * from './locations'
 export * from './modal'
