+++ client/components/AddEventForm.jsx
@@ -2,10 +2,10 @@ import React from 'react'
 import { connect } from 'react-redux'
 import * as actions from '../actions'
 import { DayPickerInput, RepeatEventForm } from './index'
+import { AddGeoSuggestInput } from './index'
 import { Field, reduxForm, SubmissionError, formValueSelector } from 'redux-form'
 import { set, get, isNil } from 'lodash'
 import moment from 'moment'
-import { GeoSuggestInput } from './index'
 
 export const renderInputField = ({ input, label, type, meta: { touched, error, warning } }) => (
     <div>
@@ -25,7 +25,7 @@ export const renderGeoSuggestInput = ({ input, label, googleApiKey, meta: { touc
     <div>
         {label && <label>{label}</label>}
         <div>
-            <GeoSuggestInput onChange={input.onChange} googleApiKey={googleApiKey} initialValue={input.value}/>
+            <AddGeoSuggestInput onChange={input.onChange} googleApiKey={googleApiKey} initialValue={input.value}/>
             {touched && ((error && <span className="help-block">{error}</span>) ||
             (warning && <span className="help-block">{warning}</span>))}
         </div>
