+++ client/components/index.js
@@ -1,5 +1,5 @@
 export { DayPickerInput } from './DayPickerInput'
-export { GeoSuggestInput } from './GeoSuggestInput'
+export { AddGeoSuggestInput } from './AddGeoSuggestInput'
 export { EventsList } from './EventsList'
 export { Event } from './Event'
 export { EventsListPanelContainer } from './EventsListPanelContainer'
