+++ server/planning/locations.py
@@ -83,7 +83,15 @@ locations_schema = {
     # Location Details
     # NewsML-G2 Event properties See:
     #    https://iptc.org/std/NewsML-G2/2.23/specification/XML-Schema-Doc-Core/ConceptItem.html#LinkC5
-    'name': {'type': 'string'},
+    #
+    # IMPORTANT: name needs to be unqiue to ensure we don't save duplicate addresses, however this is also
+    # the field where we store the formatted address, which can have variations (street number vs number street)
+    # for the same address
+    'name': {
+        'type': 'string',
+        'unique': True,
+        'mapping': not_analyzed
+    },
 
     # NewsML-G2 poiDetails properties See IPTC-G2-Implementation_Guide 12.6.3
     # or https://iptc.org/std/NewsML-G2/2.23/specification/XML-Schema-Doc-Power/ConceptItem.html#LinkAA
